requirements.txt
aniso8601==9.0.1
asttokens==2.0.5
autopep8==1.6.0
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
debugpy==1.5.1
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
feather-format==0.4.1
idna==3.3
ipykernel==6.9.2
ipython==8.1.1
jedi==0.18.1
jsonpath-ng==1.5.3
jupyter-client==7.1.2
jupyter-core==4.9.2
kaleido==0.2.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.4
numpy==1.20.2
oauthlib==3.2.0
pandas==1.4.1
parso==0.8.3
pickleshare==0.7.5
Pillow==8.2.0
plotly==5.6.0
ply==3.11
prompt-toolkit==3.0.28
psutil==5.9.0
pure-eval==0.2.2
pyarrow==7.0.0
pycodestyle==2.8.0
Pygments==2.11.2
python-dateutil==2.8.2
python-dotenv==0.19.2
python-slugify==5.0.2
pytz==2021.3
pywin32==303
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
stack-data==0.2.0
tenacity==8.0.1
text-unidecode==1.3
toml==0.10.2
tornado==6.1
traitlets==5.1.1
tweepy==4.7.0
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.0.3
